+++ src/App.tsx
@@ -1,27 +1,20 @@
+import React from 'react';
 import { useState } from 'react';
-// import '@/less/index.less';
-import 'antd/dist/antd.less';
+import { Link, Route, Routes } from 'react-router-dom';
+import ProLayout, { PageContainer } from '@ant-design/pro-layout';
 import { buildUrl } from '@/helpers/url';
-import { Button } from 'antd';
 
 function App() {
-  const [count, setCount] = useState(0);
-
   const a = buildUrl();
-  console.log(a);
 
   return (
-    <div className="App">
-      <h1>App</h1>
-
-      <div style={{ padding: '16px 0' }}>
-        <button className="btn">Button</button>
-      </div>
-
-      <div style={{ padding: '16px 0' }}>
-        <Button>Antd button</Button>
-      </div>
-    </div>
+    <ProLayout
+      style={{
+        minHeight: '100vh',
+      }}
+    >
+      <PageContainer content="this is content">Hello World</PageContainer>
+    </ProLayout>
   );
 }