+++ vite.config.ts
@@ -1,17 +1,28 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import tsconfigPaths from 'vite-tsconfig-paths';
+import vitePluginImp from 'vite-plugin-imp';
 import path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    tsconfigPaths(),
+    vitePluginImp({
+      libList: [
+        {
+          libName: 'antd',
+          style: (name) => `antd/es/${name}/style`,
+        },
+      ],
+    }),
+  ],
   server: {
     port: 5001,
   },
   resolve: {
-    alias: {
-      '@': path.resolve(__dirname, '/src'),
-    },
+    alias: [{ find: /^~/, replacement: '' }],
   },
   build: {
     outDir: 'dist',