+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 13/12/2021
+
+- Add package `vite-tsconfig-paths` to use `paths` field in tsconfig for alias import - https://github.com/aleclarson/vite-tsconfig-paths
+- [bug] Setup fix error when use `ProLayout` component
+  - Add package `vite-plugin-imp` to fix import antd 
+
 ## 12/12/2021
 
 - Add antd pro components - https://procomponents.ant.design/en-US